use dio networking manager for image upload
